+++ x/group/keeper/keeper.go
@@ -382,6 +382,15 @@ func (k Keeper) PruneProposals(ctx sdk.Context) error {
 		if err != nil {
 			return err
 		}
+		// Emit event for proposal finalized with its result
+		if err := ctx.EventManager().EmitTypedEvent(
+			&group.EventProposalPruned{
+				ProposalId:  proposal.Id,
+				Status:      proposal.Status,
+				TallyResult: &proposal.FinalTallyResult,
+			}); err != nil {
+			return err
+		}
 	}
 
 	return nil
@@ -414,6 +423,14 @@ func (k Keeper) TallyProposalsAtVPEnd(ctx sdk.Context) error {
 				return err
 			}
 			if err := k.pruneVotes(ctx, proposalID); err != nil {
+				return err
+			}
+			// Emit event for proposal finalized with its result
+			if err := ctx.EventManager().EmitTypedEvent(
+				&group.EventProposalPruned{
+					ProposalId: proposal.Id,
+					Status:     proposal.Status,
+				}); err != nil {
 				return err
 			}
 		} else if proposal.Status == group.PROPOSAL_STATUS_SUBMITTED {